Build system: Moved libcedos to its own module
